+++ src/containers/Auth/Auth.js
@@ -7,6 +7,7 @@ import Button from '../../components/UI/Button/Button';
 import Spinner from '../../components/UI/Spinner/Spinner';
 import classes from './Auth.css';
 import * as actions from '../../store/actions/index';
+import { updateObject, checkValidity } from '../../shared/utility';
 
 class Auth extends Component {
     state = {
@@ -49,48 +50,14 @@ class Auth extends Component {
         }
     }
 
-
-    checkValidity ( value, rules ) {
-        let isValid = true;
-        if ( !rules ) {
-            return true;
-        }
-
-        if ( rules.required ) {
-            isValid = value.trim() !== '' && isValid;
-        }
-
-        if ( rules.minLength ) {
-            isValid = value.length >= rules.minLength && isValid
-        }
-
-        if ( rules.maxLength ) {
-            isValid = value.length <= rules.maxLength && isValid
-        }
-
-        if ( rules.isEmail ) {
-            const pattern = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;
-            isValid = pattern.test( value ) && isValid
-        }
-
-        if ( rules.isNumeric ) {
-            const pattern = /^\d+$/;
-            isValid = pattern.test( value ) && isValid
-        }
-
-        return isValid;
-    }
-
     inputChangedHandler = ( event, controlName ) => {
-        const updatedControls = {
-            ...this.state.controls,
-            [controlName]: {
-                ...this.state.controls[controlName],
+        const updatedControls = updateObject( this.state.controls, {
+            [controlName]: updateObject( this.state.controls[controlName], {
                 value: event.target.value,
-                valid: this.checkValidity( event.target.value, this.state.controls[controlName].validation ),
+                valid: checkValidity( event.target.value, this.state.controls[controlName].validation ),
                 touched: true
-            }
-        };
+            } )
+        } );
         this.setState( { controls: updatedControls } );
     }
